+++ jtop/jetson_config.py
@@ -19,12 +19,10 @@ import os
 import curses
 # Logging
 import logging
-# Timer
-from datetime import datetime, timedelta
+
+from .gui import JTOPCONFIG
 # Create logger
 logger = logging.getLogger(__name__)
-# Gui refresh rate
-GUI_REFRESH = 1000 // 20
 # Detect user
 user = os.environ.get('USER', '')
 # Get user from sudo
@@ -40,132 +38,31 @@ JTOP_MENU = {
         (None, "Check variables status"),
         (None, "Check status permissions for {user}".format(user=user)),
         (None, "Status jetson-stats service"),
+    ],
+    'back': True
+}
+DISPLAY_MENU = {
+    'title': 'GUI menu option',
+    'menu': [
+        (None, "Text console, requiring user to login"),
+        (None, "Text console, automatically logged in as '{user}' user".format(user=user)),
+        (None, "Desktop GUI, requiring user to login"),
+        (None, "Desktop GUI, automatically logged in as '{user}' user".format(user=user)),
     ]
 }
 MAIN_PAGE = {
     'title': 'jtop {version} - main page'.format(version='AAA'),
     'menu': [
         (JTOP_MENU, "Check the status of jetson-stats"),
-        (None, "Enable/Disable boot from desktop"),
+        (DISPLAY_MENU, "Enable/Disable boot from desktop"),
         (None, "Update this tool to the latest version"),
         (None, "Information about this configuration tool"),
     ]}
 
 
-class JTOPCONFIG:
-
-    def __init__(self, stdscr):
-        # Initialize keyboard status
-        self.key = -1
-        self.old_key = -1
-        # Initialize mouse
-        self.mouse = ()
-        # Set curses reference, refresh and jetson controller
-        self.stdscr = stdscr
-        # In this program, we don't want keystrokes echoed to the console,
-        # so we run this to disable that
-        curses.noecho()
-        # Additionally, we want to make it so that the user does not have to press
-        # enter to send keys to our program, so here is how we get keys instantly
-        curses.cbreak()
-        # Try to hide the cursor
-        if hasattr(curses, 'curs_set'):
-            try:
-                curses.curs_set(0)
-            except Exception:
-                pass
-        # Lastly, keys such as the arrow keys are sent as funny escape sequences to
-        # our program. We can make curses give us nicer values (such as curses.KEY_LEFT)
-        # so it is easier on us.
-        self.stdscr.keypad(True)
-        # Enable mouse mask
-        _, _ = curses.mousemask(curses.BUTTON1_CLICKED)
-        # Refreshing page curses loop
-        # https://stackoverflow.com/questions/54409978/python-curses-refreshing-text-with-a-loop
-        self.stdscr.nodelay(1)
-        # Run loop
-        self.loop()
-
-    def draw_menu_page(self, page):
-        height, width = self.stdscr.getmaxyx()
-        # Draw menu
-        title = page['title']
-        menu = page['menu']
-        center_y = (height - len(menu)) // 2
-        # Find center on X axis
-        max_description = max([len(x) for _, x in menu])
-        center_x = (width - max_description) // 2
-        # Draw menu
-        for idx, (cmd, description) in enumerate(menu):
-            if cmd:
-                self.stdscr.addstr(center_y + idx, center_x - 5, "DO")
-            self.stdscr.addstr(center_y + idx, center_x, description)
-        # Draw title
-        title_center = (max_description - len(title)) // 2
-        self.stdscr.addstr(center_y - 2, center_x + title_center, title)
-        # Draw buttons
-        message_button = "ENTER"
-        self.stdscr.addstr(center_y + len(menu) + 1, center_x + 5, "<{message}>".format(message=message_button))
-        message_button = "ESC"
-        self.stdscr.addstr(center_y + len(menu) + 1, center_x + max_description - len(message_button) - 7, "<{message}>".format(message=message_button))
-
-    def loop(self):
-        self._current_menu = MAIN_PAGE
-        # Here is the loop of our program, we keep clearing and redrawing in this loop
-        while not self.events():
-            # First, clear the screen
-            self.stdscr.erase()
-            # Draw menu page
-            self.draw_menu_page(self._current_menu)
-            # Draw the screen
-            self.stdscr.refresh()
-            # Set a timeout and read keystroke
-            self.stdscr.timeout(GUI_REFRESH)
-
-    def events(self):
-        event = self.stdscr.getch()
-        # Run keyboard check
-        status_mouse = False
-        status_keyboard = self.keyboard(event)
-        # Clear event mouse
-        self.mouse = ()
-        # Check event mouse
-        if event == curses.KEY_MOUSE:
-            try:
-                _, mx, my, _, _ = curses.getmouse()
-                # Run event menu controller
-                #status_mouse = self.event_menu(mx, my)
-                self.mouse = (mx, my)
-            except curses.error:
-                pass
-        return status_keyboard
-
-    def keyboard(self, event):
-        self.key = event
-        if self.old_key != self.key:
-            # keyboard check list
-            if self.key == ord('q') or self.key == ord('Q') or self.ESC_BUTTON(self.key):
-                # keyboard check quit button
-                return True
-            # Store old value key
-            self.old_key = self.key
-        return False
-
-    def ESC_BUTTON(self, key):
-        """
-            Check there is another character prevent combination ALT + <OTHER CHR>
-            https://stackoverflow.com/questions/5977395/ncurses-and-esc-alt-keys
-        """
-        if key == 27:
-            n = self.stdscr.getch()
-            if n == -1:
-                return True
-        return False
-
-
 def main():
     # Run wrapper
-    curses.wrapper(JTOPCONFIG)
+    curses.wrapper(JTOPCONFIG, MAIN_PAGE)
 
 
 if __name__ == "__main__":
